requirements.txt
langchain
langchain-openai
langchain-community
langchain-pinecone
langchain-text-splitters
pinecone>=5.1.0
langfuse
ragas
openai
tiktoken
pypdf
python-docx
docx2txt
moviepy<2.0.0
openai-whisper
python-dotenv
pandas
numpy
scikit-learn
transformers
torch
pillow
fastapi
uvicorn[standard]
python-multipart
datasets
google-generativeai
langchain-google-genai
